Extract reusable calculator UI components and refactor screens
- Add `ScrollableLatexView` and `LatexDisplayCard` to standardize LaTeX expression rendering across Analysis, Calculus, and Series screens
- Extract `FocusedInputOverlay` to unify modal input focus overlays across Matrix, Distributions, and Fourier calculators
- Introduce `SegmentedToggleButtons` for mode selection and `CalculatorScreenLayout` for standardized calculator scaffolding
- Move and generalize `DataMemberBox` into `MathDataMemberBox` in `InputFieldSmall` and update cursor rendering logic
- Fix warnings project-wide and update README.md

+++ app/src/main/java/com/xemophon/aljabr/modules/graphMaker/GraphMaker.kt
@@ -202,7 +202,7 @@ fun InteractiveGraphView(expression: String) {
             val gridStep = if (viewportScale > 2f) 1f else if (viewportScale < 0.5f) 5f else 2f
 
             // Vertical grid lines
-            var startX = ((minX / gridStep).roundToInt() * gridStep).toFloat()
+            var startX = (minX / gridStep).roundToInt() * gridStep
             while (startX <= maxX) {
                 val x = centerX + startX * scaleX
                 drawLine(
